@@ -3,12 +3,18 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
-use App\User;
+use App\Models\User;
 
-class Atm extends Model{
+class Atm extends Model
+{
     protected $table = "atms";
     protected $fillable = [
-        "efectivo","session","type","motivo","user_id","branch_id"
+        "efectivo",
+        "session",
+        "type",
+        "motivo",
+        "user_id",
+        "branch_id"
     ];
     protected $hidden = [];
     protected $dates = [
@@ -16,35 +22,40 @@ class Atm extends Model{
         'created_at'
     ];
     //Relationship
-    public function user(){
+    public function user()
+    {
         return $this->belongsTo(User::class);
     }
-    public function branch(){
-        return $this->belongsTo(Config::class,'branch_id', 'id');
+    public function branch()
+    {
+        return $this->belongsTo(Config::class, 'branch_id', 'id');
     }
     //Scopes
-    public function scopeDate($query, $search){
-        if(trim($search) != ""){
-            $query->WhereDate("created_at",$search);
+    public function scopeDate($query, $search)
+    {
+        if (trim($search) != "") {
+            $query->WhereDate("created_at", $search);
         }
     }
-    public function scopeUser($query, $user, $rol){
-        if(trim($user) != ""){
+    public function scopeUser($query, $user, $rol)
+    {
+        if (trim($user) != "") {
             $user = $user * 1;
-            if(!$rol->rol && $user > 1){
-                $query->where('user_id',$user);
-            } else if($rol->rol) {
-                $query->where('user_id',$rol->id);
+            if (!$rol->rol && $user > 1) {
+                $query->where('user_id', $user);
+            } else if ($rol->rol) {
+                $query->where('user_id', $rol->id);
             }
         } else {
-            if($rol->rol){
-                $query->where('user_id',$rol->id);
+            if ($rol->rol) {
+                $query->where('user_id', $rol->id);
             }
         }
     }
-    public function scopeBranch($query, $search){
-        if(trim($search) != ""){
-            $query->where("branch_id",$search);
+    public function scopeBranch($query, $search)
+    {
+        if (trim($search) != "") {
+            $query->where("branch_id", $search);
         }
     }
 }
