@@ -4,7 +4,7 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use App\Models\Contact;
-use App\User;
+use App\Models\User;
 use App\Models\Order;
 
 class Exam extends Model
