@@ -1,12 +1,21 @@
 <?php
 
 return [
-    'paths' => ['api/*'],
+
+    'paths' => ['api/*', 'sanctum/csrf-cookie'],
+
     'allowed_methods' => ['*'],
+
     'allowed_origins' => ['*'],
+
     'allowed_origins_patterns' => [],
+
     'allowed_headers' => ['*'],
+
     'exposed_headers' => [],
+
     'max_age' => 0,
+
     'supports_credentials' => false,
+
 ];
