@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\WebController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -13,6 +14,6 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('logs', '\Rap2hpoutre\LaravelLogViewer\LogViewerController@index');
-Route::get('/', 'WebController@index')->name('/');
-Route::get('/{any}', 'WebController@index')->where('any', '.*');
+Route::get('/{any}', function () {
+    return redirect(config('app.front_url'));
+})->where('any', '.*');
