@@ -3,10 +3,10 @@
 namespace Tests\Feature;
 
 use App\Models\Order;
-use App\User;
+use App\Models\User;
 use Illuminate\Support\Str;
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Illuminate\Foundation\Testing\WithFaker;
+// use Illuminate\Foundation\Testing\RefreshDatabase;
+// use Illuminate\Foundation\Testing\WithFaker;
 use Tests\TestCase;
 
 class OrderTest extends TestCase
@@ -71,8 +71,6 @@ class OrderTest extends TestCase
         $res = $this->json('POST', 'api/orders', $data);
         dd($res->decodeResponseJson());
         $res->assertStatus(200);
-        
-        
     }
 
     public function test_get_order_show()
